[RENAME] PointDetailCustomResponse 패키지 이동
