+++ src/main/java/com/programmers/smrtstore/domain/point/infrastructure/PointDetailRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.programmers.smrtstore.domain.point.infrastructure;
 
-import com.programmers.smrtstore.domain.point.presentation.dto.res.PointDetailCustomResponse;
+import com.programmers.smrtstore.domain.point.application.dto.res.PointDetailCustomResponse;
 import java.util.List;
 
 public interface PointDetailRepositoryCustom {
